+++ README.md
@@ -1,25 +1,4 @@
 # RBT
-Project for Red and Black Tree impelementation
-
-
-//Fixes
-Ajeitei algumas coisas, mas tava pensando que talvez ficasse melhor se a gente inserisse o nó e depois fizesse as mudanças de cor e rotações. Desse jeito tá bem bagunçado
-
-//Fixes 10:11 05/10/18
-Mudança dos nomes das variáveis de objetos node para <nome>N. Ex: currentN, parentN.
-Mudança do tipo das variáveis de de node.h de pública para privada.
-Sugestão de mudança para a função tree::getParent().
-
-//Fixes 11:45 06/10/18
-Correção da função tree::getParent(), tree::getGrampa(), tree::getUncle(). 
-Mudança da função tree::insert(). 
-Criação da função tree::balance(). \n 
-Organização geral do código.
-
-// 06.10.2018 - Aparentemente a inserção tá toda coberta. Agora, precisamos fazer as remoções.
-Quando tu puder, dá uma olhada e divide elas pra gente implementar. Vou dar uma estudada em Análise agora.
-Como a gente tem até terça, dá pra fazer segunda a noite / terça de manhã tranquilo. Flwws
-
-// 07/10/18
-Arrumei os prints pra mostrar quando ocorre a inserção e quando ocorro o balanceamento, just com qual caso.
-Casos de remoção.
+Red and Black Tree implementation in C++.
+Projeto feito para a disciplina de Projeto de Algoritmos II
+do curso de Ciência da Computação da Universidade Federal do Pará.
